Remove argument in defaultTitleRenderer that isn't in use Update packages Move ResourceBrowser styling out of ResourceBrowser.tsx
