+++ frontend-web/webclient/app/Admin/Providers/Browse.tsx
@@ -90,7 +90,7 @@ function ProviderBrowse({opts}: {opts?: ResourceBrowserOpts<Provider>}): React.R
                 browser.on("fetchFilters", () => []);
 
                 browser.on("renderRow", (provider, row, dims) => {
-                    row.title.append(ResourceBrowser.defaultTitleRenderer(provider.specification.domain, dims, row));
+                    row.title.append(ResourceBrowser.defaultTitleRenderer(provider.specification.domain, row));
 
                     row.stat1.innerText = provider.owner.createdBy;
                     row.stat2.innerText = dateToString(provider.createdAt ?? timestampUnixMs());
