+++ frontend-web/webclient/package.json
@@ -21,7 +21,7 @@
         "@reduxjs/toolkit": "2.2.7",
         "@rjsf/core": "4.2.3",
         "@svgr/cli": "8.1.0",
-        "apexcharts": "3.51.0",
+        "apexcharts": "3.52.0",
         "assert": "2.1.0",
         "date-fns": "3.6.0",
         "deepcopy": "2.1.0",
@@ -30,7 +30,7 @@
         "jsrsasign": "11.1.0",
         "localforage": "1.10.0",
         "magic-bytes.js": "1.10.0",
-        "monaco-editor": "0.50.0",
+        "monaco-editor": "0.51.0",
         "react": "18.3.1",
         "react-apexcharts": "1.4.1",
         "react-datepicker": "6.9.0",
@@ -38,8 +38,8 @@
         "react-markdown": "9.0.1",
         "react-modal": "3.16.1",
         "react-redux": "8.1.3",
-        "react-router": "6.26.0",
-        "react-router-dom": "6.26.0",
+        "react-router": "6.26.1",
+        "react-router-dom": "6.26.1",
         "react-syntax-highlighter": "15.5.0",
         "react-transition-group": "4.4.5",
         "redux": "4.2.1",
@@ -52,18 +52,18 @@
         "@types/history": "5.0.0",
         "@types/json-schema": "7.0.15",
         "@types/jsrsasign": "10.5.14",
-        "@types/react": "18.3.3",
+        "@types/react": "18.3.4",
         "@types/react-datepicker": "6.2.0",
         "@types/react-dom": "18.3.0",
         "@types/react-modal": "3.16.3",
         "@types/react-redux": "7.1.33",
-        "@types/react-transition-group": "4.4.10",
+        "@types/react-transition-group": "4.4.11",
         "@types/rusha": "0.8.3",
         "@types/styled-system": "5.1.22",
         "@types/ua-parser-js": "0.7.39",
         "@vitejs/plugin-react": "4.3.1",
         "@vitejs/plugin-react-refresh": "1.3.6",
         "typescript": "5.5.4",
-        "vite": "5.4.0"
+        "vite": "5.4.2"
     }
 }
